[FIX] page objects: wizard: footer button in nested modals
Yes it's possible to have nested modal in Odoo, for example in the
preferences settings (not having multiple modals, that's a standard
feature, but having the definition of one in the middle of another).
Because of this the previous query could return the footer buttons of
the wrong one. Now we force to return the good modal-footer.

This may not work for multiple modals but let's wait for the issue to
happen.